Upgrade to phpstan 0.10.
