+++ src/Type/Spryker/AbstractSprykerDynamicTypeExtension.php
@@ -5,7 +5,7 @@
  * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
  */
 
-namespace PhpStan\Type\Spryker;
+namespace PHPStan\Type\Spryker;
 
 use Exception;
 use PhpParser\Node\Expr\MethodCall;
